+++ tests/integration/async_/test_custom_ssl_context.py
@@ -38,12 +38,14 @@ async def test_custom_ssl_context_wraps_connection(uri, auth, mocker):
     fake_ssl_context.wrap_socket.side_effect = wrap_fail
     fake_ssl_context.wrap_bio.side_effect = wrap_fail
 
-    async with neo4j.AsyncGraphDatabase.driver(
-        uri, auth=auth, ssl_context=fake_ssl_context
-    ) as driver:
-        async with driver.session() as session:
-            with pytest.raises(NoNeedToGoFurtherError):
-                await session.run("RETURN 1")
+    async with (
+        neo4j.AsyncGraphDatabase.driver(
+            uri, auth=auth, ssl_context=fake_ssl_context
+        ) as driver,
+        driver.session() as session,
+    ):
+        with pytest.raises(NoNeedToGoFurtherError):
+            await session.run("RETURN 1")
 
     assert (
         fake_ssl_context.wrap_socket.call_count