+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 See also https://github.com/neo4j/neo4j-python-driver/wiki for a full changelog.
 
 ## NEXT RELEASE
+- Python 3.7, 3.8, and 3.9 support has been dropped.
 - Changed errors raised under certain circumstances
   - `neo4j.exceptions.UnsupportedServerProduct` if no common bolt protocol version could be negotiated with the server
    (instead of internal `neo4j._exceptions.BoltHandshakeError`).